requirements.txt
anthropic
tavily-python
pypdf
pymupdf
python-pptx
python-dotenv
requests
firecrawl-py
google-genai>=1.0.0
